Adding csv files for the data worked on Tuesday 6/20
For the overview, Isabele listed all of the buildings, Erika put them in alphabetical order. Multiple members of the team worked on the overview for tags for Room Types and Edges. For Amos Eaton, Isable put in the rooms for the second and third floors. Erika did the first and fourth floors.