Mapping the first floor of Amos Eaton to a cartesian plane.
I imported a png file of the floor plan for AE 100 level into desmos, so fit it onto
a coordinate plane. I added lines and points to symbolize hallways and doorways to rooms.
This commit contains a png file with the lines, points, and cartesian plane on it,
as well as a text fle listing all of the lines and points that were put in. I
also documented the process as a reminder for future use.