Updated 87 Gym spreadsheet according to the datbase model
Reformatted the spreadsheet to be compatible with importing directly into the database.
Added the third floor since it was missing, also fixed the Bathroom attributesto include gender.
Made sure all the attributes present in the database room relation are also in the spreadsheet
and in the correct order. Made the room types uniform (ex: changed 'Hall' and 'Hallway' to all be 'Hallway').
Also made all room types capitalized.