Documentation for what I did on 7/21
Three documentation files:
data_to_collect is the start of a list of data that we will need to get, potentially having others help to get it.
data_base_updates discusses some changes we made to the database design and some decisions we made regarding the design
I also updated the making_database file with updates on our progress with importing the csv data into the database.
for the importing data, long story short, we're still trying to figure out how to do it.