Inputting Data into the Database from csv files with csv files reformatted accordingly
I documented the useful information we came across while figuring out how to do this whole thing
in the making_database.txt file. 87 Gym and DCC csv files are reformatted to
match how they need to be formatted for being imported into the database.
I included the lines of code in the sql file that will input data for 87 gym, amos eaton, and DCC
into the database.