Fixed/Updated Spreadsheets for Lally, Pittsburgh, Ricketts, West
I reformatted the spreadsheets for these four buildings in accordance with the database model.
I made sure all room types correctly corresponded to the options presented in the database model.
I added {0,0} for all of the door coordinates so that the attribute isn't empty, since it is a primary
attribute it cannot be null. I updated room name for the few that I knew in these buildings.