Cartesian Mapping of Second floor Amos Eaton
I imposed the floor plan for the second floor of amos eaton onto a cartesian plane.
I used lines for hallways and points for doorways in the hallways. We are not
allowed to post the image of floor plans online, so I also recorded the lines and points
that I marked down. I also inputted this data into the csv file.