cartesian mapping amos eaton 300 and 400 levels
Mapped the 3rd and 4th floors of Amos Eaton to the Cartesian plane.
Similarly, I cannot upload the image of the mapping due to not
being allowed to post the floor plans online, but I have included a text file
that describes all lines and points in each mapping. I used this data to update the
csv file as well.